+++ lib/src/command/enum/enum.dart
@@ -1,3 +1,3 @@
+export 'enum_start.dart';
 export 'models/enum_common.dart';
 export 'models/enum_item.dart';
-export 'enum_start.dart';