+++ lib/src/constants.dart
@@ -17,7 +17,7 @@ class ConstHelper {
 
   static const String configName = 'turn_gen';
   static const String packageOnlyLetter = 'turngen';
-  static const String version = '0.7.1';
+  static const String version = '0.8.0';
 }
 
 class ConstConsole {
@@ -73,7 +73,6 @@ class MyRegexText {
     return RegExp(commentSpaceEnd, multiLine: true);
   }
 
-  
   static RegExp comment() {
     return RegExp(spaceCommentSpace, multiLine: true);
   }